added conversion and poll "hook"

+++ TelegramNeuralServerAPI/source/BotControl/UserHandler/LocalImage.cs
@@ -7,12 +7,12 @@ using System.Threading.Tasks;
 
 namespace TelegramNeuralServerAPI
 {
-	internal class LocalImage(byte width, byte height, byte channels, string data)
+	internal class LocalImage(ushort width, ushort height, byte channels, string data)
 	{
 		[JsonInclude]
-		public byte width = width;
+		public ushort width = width;
 		[JsonInclude]
-		public byte height = height;
+		public ushort height = height;
 		[JsonInclude]
 		public byte channels = channels;
 		[JsonInclude]
